se corrigio integer por long en repository se añadio funciones de hotelservice falta agregar enum de caracteristicas de habitacion

+++ src/main/java/com/example/demo/repository/HabitacionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface HabitacionRepository extends JpaRepository<Habitacion,Integer> {
+public interface HabitacionRepository extends JpaRepository<Habitacion,Long> {
     List<Habitacion> findByUsuario_Id(Integer usuarioId);
 
 }
